requirements.txt
pyserial
numpy
pygame
pyqt5